backend/requirements.txt
fastapi
uvicorn[standard]
google-generativeai
python-multipart
pydantic
pydantic-settings
python-dotenv
